Refractor stuff from form.py
