+++ src/TOTKOptimizer.ini
@@ -1,5 +1,5 @@
 [Time]
-api = 1709019212.787935
+api = 1709026505.6603122
 
 [Updates]
 dfps = UltraCam
