--- src/modules/json.data/Description.json
@@ -1,1 +1,0 @@
-{"HARDCODED MODS - GROUP": "None", "Resolution": "Choose the desired resolution for Tears of The Kingdom.\nNOTE: This is an internal upscaler, using Dynamic FPS.\nExtended Memory layout in Yuzu is Mandatory.", "FPS": "Choose the desired FPS LIMIT for Tears of The Kingdom.", "Shadows": "Choose the desired SHADOW resolution for Tears of The Kingdom.\nNote: Recommended to use 1024x.", "Camera Quality": "Increases camera Picture Quality.\nNote: Recommended to not use this mod, as it COULD cause issues.", "UI": "Choose the button prompts for Tears of The Kingdom.\nNote: Recommended to use BlackscreenFIX, if you don't desire to use any other UI mod.", "First Person": "Switches all cameras to First Person.", "PRESETS - GROUP": "None", "Presets": "Preset Mods for Tears of The Kingdom.", "Settings": "Select Yuzu PRESET settings, based on your PC configuration.", "BUTTONS - GROUP": "None", "Reset": "Removes the portable path and defaults to Appdata", "Browse": "Search for PORTABLE Yuzu/Ryujinx.exe.", "Backup": "Backups your Tears of The Kingdom saves.", "Kofi": "If you want to support this project.\nfeel free to check out my Kofi page.", "Github": "Leads to the Project page.", "Apply": "Applies the selected settings!", "Switch": "Switches between Yuzu and Ryujinx Manager.", "Main": "The Main section of the Mod Manager.", "Cheats": "In this tab you can add cheats to your game.", "DYNAMICMODS - GROUP": "None", "Remove QuaReduction": "Improves Quality when FPS drops.\nThis is mandatory with new DFPS.", "Remove DOF": "Removes Depth of Field effect.", "Remove Lensflare": "Removes lensflare effect,\nNote: The lensflare in TOTK flickers a lot..", "Improve LOD": "Improves the distant objects in the game.\nThis has no effect on performance.", "Camera Speed x2": "Increase the Player Camera Speed.", "Remove FXAA": "Removes Internal FX Anti Aliasing, \nNOTE: could improve quality with external upscaler.", "Remove FSR": "Removes Internal FSR 1.0 Upscaler, \nNOTE: could reduce blurriness and improve quality.", "CHEATS - GROUP": "None", "God Mode": "Turn on God Mode", "Mark all Korok": "Marks all the korok seeds on the map \nNote: Zoom in!.", "Free Autobuild": "Removes the cost of Autobuild.", "Infinite Battery": "Zonai devices consume no battery.", "Infinite Durability": "Infinite Durability for Weapons and Shields.", "Infinite Items (500)": "Makes all items Infinite,\nNote: recommended to do a backup.", "Infinite Zonai Devices": "Lets you summon Zonai devices infinitely.", "Infinite Stamina": "Infinite running Stamina.", "Swimming Speed 2x": "Increases Swimming speed by 200%.", "Swimming Speed 3x": "Increases Swimming speed by 300%.", "Fast Forward Time": "Press (ZL+R3Left) to accelerate time.", "Time Resume": "Press (ZL+R3Up) to trigger.", "Time Rewind": "Time Rewind press (ZL+R3Left) to trigger", "Time Stop": "Press (ZL+R3Down) to trigger.", "Mark ALl Shrines": "Marks all the shrines on the map.", "Quickfire Bow": "Decreases bow Charge time.", "No Charge Stamina": "Removes weapon charge stamina.", "No Fall Damage": "Removes fall damage.", "Movement speed 3x": "Movement speed 300%, \nNOTE: Affects horses.", "Movement speed 2x": "Movement speed 200%, \nNOTE: Affects horses.", "Movement speed 1x": "Movement speed 100%, \nNOTE: Affects horses.", "No skill CD": "Removes the skill CD of sages."}
