--- src/modules/json.data/Version.json
@@ -1,1 +1,0 @@
-[{"version": "main1.0.0", "Source": "\nFile Generated by @MaxLastBreath, original source: https://github.com/MaxLastBreath/TOTK-mods", "nsobid": "\n@nsobid-082CE09B06E33A123CB1E2770F5F9147709033DB", "offset": "\n@flag print_values\n@flag offset_shift 0x100\n", "Camera Speed x2": "@enabled\n03775f6c 0000A03f\n00d0531c 09902E1E\n@stop", "Remove QuaReduction": "@enabled\n00C40A8C C0000014\n@stop\n", "Remove DOF": "@enabled\n00BD0F14 C0035FD6\n@stop\n", "Remove Lensflare": "@enabled\n029ca250 1F2003D5\n@stop\n", "Improve LOD": "\\\\Enabled by default on 1.0.0\n", "Remove FXAA": "@enabled\n00C40A70 08008052\n@stop\n", "Remove FSR": "@enabled\n00C40A80 08008052\n@stop\n"}, {"version": "main1.1.0", "Source": "File Generated by @MaxLastBreath, original source: https://github.com/MaxLastBreath/TOTK-mods", "nsobid": "\n@nsobid-D5AD6AC71EF53E3E52417C1B81DBC9B4142AA3B3", "offset": "\n@flag print_values\n@flag offset_shift 0x100\n", "Camera Speed x2": "@enabled\n0381648C 0000A03f\n00D228DC 09902E1E\n@stop", "Remove QuaReduction": "@enabled\n00CAD34C C0000014\n@stop\n", "Remove DOF": "@enabled\n00C488B4 C0035FD6\n@stop\n", "Remove Lensflare": "@enabled\n02A42490 1F2003D5\n@stop\n", "Improve LOD": "@enabled\n027D9448 24000014\n@stop\n", "Remove FXAA": "@enabled\n00CAD330 08008052\n@stop\n", "Remove FSR": "@enabled\n00CAD340 08008052\n@stop\n"}, {"version": "main1.1.1", "Source": "\nFile Generated by @MaxLastBreath, original source: https://github.com/MaxLastBreath/TOTK-mods", "nsobid": "\n@nsobid-168DD518D925C7A327677286E72FEDA833314919", "offset": "\n@flag print_values\n@flag offset_shift 0x100\n", "Camera Speed x2": "@enabled\n0381f26c 0000A03f\n00d01e4c 09902E1E\n@stop", "Remove QuaReduction": "@enabled\n00CC1C2C C0000014\n@stop\n", "Remove DOF": "@enabled\n00C25898 C0035FD6\n@stop\n", "Remove Lensflare": "@enabled\n02A4A1E0 1F2003D5\n@stop\n", "Improve LOD": "@enabled\n027E07F8 10000014\n@stop\n", "Remove FXAA": "@enabled\n00CC1C10 08008052\n@stop\n", "Remove FSR": "@enabled\n00CC1C20 08008052\n@stop\n"}, {"version": "main1.1.2", "Source": "\nFile Generated by @MaxLastBreath, original source: https://github.com/MaxLastBreath/TOTK-mods", "nsobid": "\n@nsobid-9A10ED9435C06733DA597D8094D9000AB5D3EE60", "offset": "\n@flag print_values\n@flag offset_shift 0x100\n", "Camera Speed x2": "@enabled\n0380ee6c 0000A03f\n00d1411c 09902E1E\n@stop", "Remove QuaReduction": "@enabled\n00C76FFC C0000014\n@stop\n", "Remove DOF": "@enabled\n00C4B934 C0035FD6\n@stop\n", "Remove Lensflare": "@enabled\n02A395F0 1F2003D5\n@stop\n", "Improve LOD": "@enabled\n027D04F8 24000014\n@stop\n", "Remove FXAA": "@enabled\n00C76FE0 08008052\n@stop\n", "Remove FSR": "@enabled\n00C76FF0 08008052\n@stop\n"}, {"version": "main1.2.0", "Source": "\nFile Generated by @MaxLastBreath, original source: https://github.com/MaxLastBreath/TOTK-mods", "nsobid": "\n@nsobid-6F32C68DD3BC7D77AA714B80E92A096A737CDA77", "offset": "\n@flag print_values\n@flag offset_shift 0x100\n", "Camera Speed x2": "@enabled\n03802acc 0000A03F\n00ce3834 09902E1E\n@stop\n", "Remove QuaReduction": "@enabled\n00c4275c C0000014\n@stop\n", "Remove DOF": "@enabled\n00c07024 C0035FD6\n@stop\n", "Remove Lensflare": "@enabled\n02a2d460 1F2003D5\n@stop\n", "Improve LOD": "@enabled\n027c3ea8 24000014\n@stop\n", "Remove FXAA": "@enabled\n00c42740 08008052\n@stop\n", "Remove FSR": "@enabled\n00c42750 08008052\n@stop\n"}, {"version": "main1.2.1", "Source": "\nFile Generated by @MaxLastBreath, original source: https://github.com/MaxLastBreath/TOTK-mods", "nsobid": "\n@nsobid-9B4E43650501A4D4489B4BBFDB740F26AF3CF850", "offset": "\n@flag print_values\n@flag offset_shift 0x100\n", "Camera Speed x2": "@enabled\n03837390 0000A03F\n00D180F8 09902E1E\n@stop\n", "Remove QuaReduction": "@enabled\n00C7EB4C C0000014\n@stop\n", "Remove DOF": "@enabled\n00C5B128 C0035FD6\n@stop\n", "Remove Lensflare": "@enabled\n02A3C7B0 1F2003D5\n@stop\n", "Improve LOD": "@enabled\n027D2268 24000014\n@stop\n", "Remove FXAA": "@enabled\n00C7EB30 08008052\n@stop\n", "Remove FSR": "@enabled\n00C7EB40 08008052\n@stop\n"}]
