--- src/modules/json.data/preset.json
@@ -1,1 +1,0 @@
-{"Default": {"Resolution": "1080p FHD", "FPS": "60", "ShadowResolution": "High x1024", "CameraQuality": "Disable", "UI": "Black Screen Fix", "First Person": "Disabled", "Remove QuaReduction": "Enable", "Remove DOF": "Enable", "Remove Lensflare": "Enable", "Improve LOD": "Enable", "Remove FSR": "Disable", "Remove FXAA": "Disable"}, "Low": {"Resolution": "1080p FHD", "FPS": "30", "ShadowResolution": "Medium x512", "CameraQuality": "Disable", "UI": "Black Screen Fix", "First Person": "Disabled", "Remove QuaReduction": "Disable", "Remove DOF": "Enable", "Remove Lensflare": "Enable", "Improve LOD": "Enable", "Remove FSR": "Disable", "Remove FXAA": "Disable"}, "Medium": {"Resolution": "1440p QHD - Experimental", "FPS": "30", "ShadowResolution": "High x1024", "CameraQuality": "Disable", "First Person": "Disabled", "UI": "Black Screen Fix", "Remove QuaReduction": "Enable", "Remove DOF": "Disable", "Remove Lensflare": "Enable", "Improve LOD": "Enable", "Remove FSR": "Enable", "Remove FXAA": "Disable"}, "High": {"Resolution": "4K UHD", "FPS": "60", "ShadowResolution": "High x1024", "CameraQuality": "Disable", "First Person": "Disabled", "UI": "Black Screen Fix", "Remove QuaReduction": "Enable", "Remove DOF": "Enable", "Remove Lensflare": "Enable", "Improve LOD": "Enable", "Remove FSR": "Enable", "Remove FXAA": "Enable"}, "Ultra": {"Resolution": "6K QUHD - Experimental", "FPS": "60", "ShadowResolution": "High x1024", "CameraQuality": "Disable", "First Person": "Disabled", "UI": "Black Screen Fix", "Remove QuaReduction": "Enable", "Remove DOF": "Enable", "Remove Lensflare": "Enable", "Improve LOD": "Enable", "Remove FSR": "Enable", "Remove FXAA": "Enable"}, "Steamdeck": {"Resolution": "1080p FHD", "FPS": "30", "ShadowResolution": "Medium x512", "CameraQuality": "Disable", "First Person": "Disabled", "UI": "Black Screen Fix", "Remove QuaReduction": "Disable", "Remove DOF": "Enable", "Remove Lensflare": "Enable", "Improve LOD": "Enable", "Remove FSR": "Enable", "Remove FXAA": "Disable"}, "RTX 4090 SUPER HIGH": {"Resolution": "6K QUHD - Experimental", "FPS": "60", "ShadowResolution": "High x1024", "CameraQuality": "Disable", "First Person": "Disabled", "UI": "Black Screen Fix", "Remove QuaReduction": "Enable", "Remove DOF": "Enable", "Remove Lensflare": "Enable", "Improve LOD": "Enable", "Remove FSR": "Enable", "Remove FXAA": "Enable"}}
