+++ src/modules/modules.py
@@ -1,0 +1,15 @@
+from modules.canvas import *
+from modules.qt_config import modify_disabled_key, get_config_parser
+from modules.checkpath import checkpath, DetectOS
+from modules.backup import *
+from modules.logger import *
+from modules.launch import *
+from modules.config import *
+from modules.util import *
+from modules.patches import *
+from modules.image_paths import *
+from modules.presets import *
+from modules.load_elements import *
+from modules.benchmarks import *
+
+# custom imports
